Move ForegroundStyle to separate file
